Move categories entity to /assignments
