+++ test/activities/assignments/CategoriesEntity.html
@@ -10,11 +10,10 @@
 		<script src="/node_modules/sinon/pkg/sinon.js"></script>
 		<script type="module" src="/node_modules/fetch-mock/es5/client-bundle.js"></script>
 
-		<script type="module" src="../../../siren-parser/global.js"></script>
-		<script type="module" src="../../src/activities/CategoriesEntity.js"></script>
+		<script type="module" src="../../../../siren-parser/global.js"></script>
 	</head>
-
 	<body>
 		<script type="module" src="./CategoriesEntity.js"></script>
 	</body>
 </html>
+
