+++ test/activities/assignments/data/EditableCategories.js
