+++ test/activities/assignments/data/EmptyCategories.js
