+++ test/activities/assignments/data/NonSelectedCategories.js
