+++ test/index.html
@@ -37,7 +37,7 @@
 			'./activities/content/ContentLTILinkEntity.html',
 			'./activities/assignments/AssignmentActivityUsageEntity.html',
 			'./activities/assignments/AssignmentEntity.html',
-			'./activities/CategoriesEntity.html',
+			'./activities/assignments/CategoriesEntity.html',
 			'./activities/quizzes/QuizEntity.html',
 			'./activities/quizzes/attempts/QuizAttemptsEntity.html',
 			'./activities/quizzes/timing/QuizTimingEntity.html',
